[cli] Use local context in `werft job get`
Fixes 32leaves/werft#31
