+++ cmd/client/job-logs.go
@@ -41,28 +41,19 @@ var jobLogsCmd = &cobra.Command{
 		client := v1.NewWerftServiceClient(conn)
 		ctx := context.Background()
 
-		var name string
+		var (
+			name string
+			err  error
+		)
 		if len(args) == 0 {
-			filter, err := getLocalContextJobFilter()
+			name, err = findJobByLocalContext(ctx, client)
 			if err != nil {
 				return err
 			}
-
-			resp, err := client.ListJobs(ctx, &v1.ListJobsRequest{
-				Filter: filter,
-				Order: []*v1.OrderExpression{&v1.OrderExpression{
-					Field:     "created",
-					Ascending: false,
-				}},
-			})
-			if err != nil {
-				return err
-			}
-			if len(resp.Result) == 0 {
+			if name == "" {
 				return xerrors.Errorf("no job found - please specify job name")
 			}
 
-			name = resp.Result[0].Name
 			fmt.Printf("showing logs of \033[34m\033[1m%s\t\033\033[0m\n", name)
 		} else {
 			name = args[0]
